调整级联目录
