--- include/aquarius/stream.hpp
@@ -1,7 +1,0 @@
-#pragma once
-#include "stream/streambuf.hpp"
-
-namespace aquarius
-{
-	using ftstream = message_streambuf<>;
-}
